+++ src/Pricing/Assurant.Pricing.Domain/RuleEngine/RuleEngine.cs
@@ -9,7 +9,7 @@ namespace Assurant.Pricing.Domain.RuleEngine
 {
     public abstract class RuleEngine : IRuleEngine
     {
-        public abstract double CalculatePrice(ITicket ticket, IPriceRepository priceRepository, IHolidayRepository holidayRepository);
+        public abstract decimal CalculatePrice(ITicket ticket, IPriceRepository priceRepository, IHolidayRepository holidayRepository);
 
         public abstract void SetComponent(IRuleEngine ruleEngine);
